+++ docs/quantization_weight_only.md
@@ -124,11 +124,11 @@ To find the best algorithm, users can leverage the `autotune` feature to explore
 ### **User code example**
 
 ```python
-from onnx_neural_compressor.quantization import get_woq_tuning_config, autotune
-from onnx_neural_compressor.utils.base_tuning import TuningConfig
+from onnx_neural_compressor import config
+from onnx_neural_compressor.quantization import tuning
 
-tune_config = TuningConfig(config_set=get_woq_tuning_config())
-best_model = autotune(
+tune_config = tuning.TuningConfig(config_set=config.get_woq_tuning_config())
+best_model = tuning.autotune(
     model_input=model,
     tune_config=tune_config,
     eval_fn=eval_fn,
