+++ examples/nlp/huggingface_model/text_generation/llama/quantization/weight_only/requirements.txt
@@ -7,6 +7,6 @@ onnxruntime-extensions; python_version < '3.11'
 datasets
 optimum
 evaluate
-intel-extension-for-transformers
+intel-extension-for-transformers >= 1.4.1
 peft
-git+https://github.com/EleutherAI/lm-evaluation-harness.git@cc9778fbe4fa1a709be2abed9deb6180fd40e7e2
+lm-eval==0.4.2
