+++ onnx_neural_compressor/algorithms/layer_wise/__init__.py
@@ -11,7 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from onnx_neural_compressor.algorithms.layer_wise.core import layer_wise_quant
-
-__all__ = ["layer_wise_quant"]
