+++ onnx_neural_compressor/algorithms/weight_only/rtn.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
 # Copyright (c) 2023 MIT HAN Lab
@@ -17,31 +18,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import os
-from pathlib import Path
-from typing import List, Union
+import pathlib
 
 import numpy as np
 import onnx
 import onnxruntime as ort
-from packaging.version import Version
+from onnx_neural_compressor import config
+from onnx_neural_compressor import constants
+from onnx_neural_compressor import onnx_model
+from onnx_neural_compressor import utility
+from onnx_neural_compressor.algorithms.layer_wise import core
+from onnx_neural_compressor.algorithms.weight_only import utility as woq_utility
+from packaging import version
 
-from onnx_neural_compressor.algorithms.weight_only.utility import (
-    make_matmul_weight_only_node,
-    pad_tensor,
-    qdq_tensor,
-    quant_tensor,
-)
-from onnx_neural_compressor.quantization.config import RTNConfig
-from onnx_neural_compressor.utils import ONNXRT116_VERSION, ONNXRT1161_VERSION, dtype_mapping, simple_progress_bar
-from onnx_neural_compressor.utils.onnx_model import ONNXModel
-
-__all__ = ["apply_rtn_on_model", "rtn_quantize"]
+from typing import List, Union  # isort: skip
 
 
 def rtn_quantize(
-    model: Union[onnx.ModelProto, ONNXModel, Path, str],
+    model: Union[onnx.ModelProto, onnx_model.ONNXModel, pathlib.Path, str],
     weight_config: dict = {},
     num_bits: int = 4,
     group_size: int = 32,
@@ -54,7 +49,7 @@ def rtn_quantize(
     """Quantize the model with round to nearst method.
 
     Args:
-        model (Union[onnx.ModelProto, ONNXModel, Path, str]): onnx model
+        model (Union[onnx.ModelProto, onnx_model.ONNXModel, pathlib.Path, str]): onnx model
         weight_config (dict, optional): quantization config
             For example,
             weight_config = {
@@ -81,8 +76,8 @@ def rtn_quantize(
     Returns:
         onnx.ModelProto: quantized onnx model.
     """
-    if not isinstance(model, ONNXModel):
-        model = ONNXModel(model)
+    if not isinstance(model, onnx_model.ONNXModel):
+        model = onnx_model.ONNXModel(model)
     base_dir = os.path.dirname(model.model_path) if model.model_path is not None else ""
     new_nodes = []
     remove_nodes = []
@@ -91,7 +86,7 @@ def rtn_quantize(
     for node in model.nodes():
         if node.op_type in ["MatMul"]:
             curr_id += 1
-            simple_progress_bar(total_num, curr_id)
+            utility.simple_progress_bar(total_num, curr_id)
 
         # check op_type of node is MatMul
         # check dim 1 of input is weight tensor
@@ -119,11 +114,13 @@ def rtn_quantize(
             k_blocks = (org_w_shape[0] - 1) // group_size + 1
             init_share_num = model.get_initializer_share_num(node.input[1])
 
-            weight = pad_tensor(weight, group_size, k_blocks)
+            weight = woq_utility.pad_tensor(weight, group_size, k_blocks)
 
-            satisfy_MatMulNBits_condition = Version(ort.__version__) > ONNXRT1161_VERSION and num_bits == 4
+            satisfy_MatMulNBits_condition = (
+                version.Version(ort.__version__) > constants.ONNXRT1161_VERSION and num_bits == 4
+            )
             satisfy_MatMulFpQ4_condition = (
-                Version(ort.__version__) >= ONNXRT116_VERSION and num_bits == 4 and group_size == 32
+                version.Version(ort.__version__) >= constants.ONNXRT116_VERSION and num_bits == 4 and group_size == 32
             )
             if ("CUDAExecutionProvider" in providers and satisfy_MatMulNBits_condition) or (
                 "CUDAExecutionProvider" not in providers
@@ -131,10 +128,10 @@ def rtn_quantize(
             ):  # pragma: no cover
                 # MatMulFpQ4 support 4 bits and 32 group_size with ort 1.16.0 and 1.16.1 versions, supported by CPU EP
                 # MatMulNBits supports 4 bits and 2^n group_size with ort > 1.16.1, supported by CPU EP AND CUDA EP
-                q_weight, scale, zp = quant_tensor(
+                q_weight, scale, zp = woq_utility.quant_tensor(
                     weight.T, num_bits, group_size, scheme, "uint", ratios.get(node.input[1], 1)
                 )
-                q_matmul_node, new_inits = make_matmul_weight_only_node(
+                q_matmul_node, new_inits = woq_utility.make_matmul_weight_only_node(
                     node=node,
                     weight_shape=org_w_shape,
                     num_bits=num_bits,
@@ -150,13 +147,15 @@ def rtn_quantize(
                 remove_nodes.append(node)
                 new_nodes.append(q_matmul_node)
             else:
-                q_weight = qdq_tensor(weight.T, num_bits, group_size, scheme, "int", ratios.get(node.input[1], 1))
+                q_weight = woq_utility.qdq_tensor(
+                    weight.T, num_bits, group_size, scheme, "int", ratios.get(node.input[1], 1)
+                )
                 q_weight = np.reshape(q_weight, (org_w_shape[1], -1))
                 q_weight = np.transpose(q_weight)
                 q_weight = q_weight[: org_w_shape[0], :].astype(dtype)
                 q_weight_tensor = onnx.helper.make_tensor(
                     name=node.input[1] + "_Q{}G{}".format(str(num_bits), str(group_size)),
-                    data_type=dtype_mapping[str(dtype)],
+                    data_type=utility.dtype_mapping[str(dtype)],
                     dims=weight.shape,
                     vals=q_weight.tobytes(),
                     raw=True,
@@ -172,9 +171,7 @@ def rtn_quantize(
 
     # reload external data to prevent external data file path errors
     if model.is_large_model:
-        from onnx.external_data_helper import load_external_data_for_model
-
-        load_external_data_for_model(model.model, os.path.split(model.model_path)[0])
+        onnx.external_data_helper.load_external_data_for_model(model.model, os.path.split(model.model_path)[0])
 
     if return_modelproto:
         return model.model
@@ -182,11 +179,13 @@ def rtn_quantize(
         return model
 
 
-def apply_rtn_on_model(model: Union[onnx.ModelProto, ONNXModel, Path, str], quant_config: dict) -> onnx.ModelProto:
+def apply_rtn_on_model(
+    model: Union[onnx.ModelProto, onnx_model.ONNXModel, pathlib.Path, str], quant_config: dict
+) -> onnx.ModelProto:
     """Apply RTN on onnx model.
 
     Args:
-        model (Union[onnx.ModelProto, ONNXModel, Path, str]): onnx model.
+        model (Union[onnx.ModelProto, onnx_model.ONNXModel, pathlib.Path, str]): onnx model.
         quant_config (dict): quantization config.
 
     Returns:
@@ -197,20 +196,20 @@ def apply_rtn_on_model(model: Union[onnx.ModelProto, ONNXModel, Path, str], quan
 
     # set other model params
     quant_kwargs = {}
-    quant_kwargs = {key: quant_config.pop(key) for key in RTNConfig.model_params_list if key in quant_config}
+    quant_kwargs = {key: quant_config.pop(key) for key in config.RTNConfig.model_params_list if key in quant_config}
 
     # change op config to dict type
     for op_name_type, op_config in quant_config.items():
-        if isinstance(op_config, RTNConfig):
+        if isinstance(op_config, config.RTNConfig):
             quant_config[op_name_type] = op_config.to_dict()
 
     if layer_wise:
-        from onnx_neural_compressor.algorithms import layer_wise_quant
-
-        quantized_model = layer_wise_quant(model, quant_func=rtn_quantize, weight_config=quant_config, **quant_kwargs)
+        quantized_model = core.layer_wise_quant(
+            model, quant_func=rtn_quantize, weight_config=quant_config, **quant_kwargs
+        )
     else:
         quantized_model = rtn_quantize(model, weight_config=quant_config, **quant_kwargs)
 
-    if isinstance(quantized_model, ONNXModel):
+    if isinstance(quantized_model, onnx_model.ONNXModel):
         quantized_model = quantized_model.model
     return quantized_model
