+++ onnx_neural_compressor/quantization/__init__.py
@@ -12,65 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from onnx_neural_compressor.quantization.algorithm_entry import (
-    smooth_quant_entry,
-    rtn_quantize_entry,
-    gptq_quantize_entry,
-    awq_quantize_entry,
-)
-from onnx_neural_compressor.quantization.calibrate import CalibrationDataReader
-from onnx_neural_compressor.quantization.config import (
-    RTNConfig,
-    get_default_rtn_config,
-    GPTQConfig,
-    get_default_gptq_config,
-    AWQConfig,
-    get_default_awq_config,
-    SmoothQuantConfig,
-    get_default_sq_config,
-    StaticQuantConfig,
-    DynamicQuantConfig,
-    get_woq_tuning_config,
-)
-from onnx_neural_compressor.quantization.autotune import autotune, get_all_config_set
 
-from onnx_neural_compressor.quantization.matmul_4bits_quantizer import MatMul4BitsQuantizer
-from onnx_neural_compressor.quantization.matmul_nbits_quantizer import (
-    RTNWeightOnlyQuantConfig,
-    GPTQWeightOnlyQuantConfig,
-    AWQWeightOnlyQuantConfig,
-    MatMulNBitsQuantizer,
-)
-
+from onnxruntime.quantization.quant_utils import QuantFormat
+from onnxruntime.quantization.quant_utils import QuantType
 from onnx_neural_compressor.quantization.quantize import quantize
-
-from onnxruntime.quantization.quant_utils import QuantFormat, QuantType
-
-__all__ = [
-    "smooth_quant_entry",
-    "rtn_quantize_entry",
-    "gptq_quantize_entry",
-    "awq_quantize_entry",
-    "RTNConfig",
-    "get_default_rtn_config",
-    "GPTQConfig",
-    "get_default_gptq_config",
-    "AWQConfig",
-    "get_default_awq_config",
-    "SmoothQuantConfig",
-    "get_default_sq_config",
-    "get_woq_tuning_config",
-    "get_all_config_set",
-    "StaticQuantConfig",
-    "DynamicQuantConfig",
-    "CalibrationDataReader",
-    "autotune",
-    "MatMul4BitsQuantizer",
-    "MatMulNBitsQuantizer",
-    "RTNWeightOnlyQuantConfig",
-    "GPTQWeightOnlyQuantConfig",
-    "AWQWeightOnlyQuantConfig",
-    "quantize",
-    "QuantFormat",
-    "QuantType",
-]
