+++ onnx_neural_compressor/quantization/algorithm_entry.py
@@ -12,46 +12,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pathlib
 import tempfile
-from pathlib import Path
 from typing import Union
 
 import onnx
-from onnxruntime.quantization import quantize
-
-from onnx_neural_compressor.algorithms import Smoother
-from onnx_neural_compressor.quantization.calibrate import CalibrationDataReader
-from onnx_neural_compressor.quantization.config import AWQConfig, GPTQConfig, RTNConfig, SmoothQuantConfig
-from onnx_neural_compressor.utils import AWQ, GPTQ, RTN, SMOOTH_QUANT, logger
-from onnx_neural_compressor.utils.utility import register_algo
-
-__all__ = [
-    "smooth_quant_entry",
-    "rtn_quantize_entry",
-    "gptq_quantize_entry",
-    "awq_quantize_entry",
-]
+from onnx_neural_compressor import config
+from onnx_neural_compressor import constants
+from onnx_neural_compressor import data_reader
+from onnx_neural_compressor import logger
+from onnx_neural_compressor import utility
+from onnx_neural_compressor.algorithms.smoother import core
+from onnx_neural_compressor.algorithms.weight_only import awq
+from onnx_neural_compressor.algorithms.weight_only import gptq
+from onnx_neural_compressor.algorithms.weight_only import rtn
+from onnxruntime import quantization
 
 
 ###################### SmoothQuant Entry ##################################
-@register_algo(name=SMOOTH_QUANT)
+@utility.register_algo(name=constants.SMOOTH_QUANT)
 def smooth_quant_entry(
-    model: Union[Path, str],
-    quant_config: SmoothQuantConfig,
-    calibration_data_reader: CalibrationDataReader,
-    model_output: Union[Path, str] = None,
+    model: Union[pathlib.Path, str],
+    quant_config: config.SmoothQuantConfig,
+    calibration_data_reader: data_reader.CalibrationDataReader,
+    model_output: Union[pathlib.Path, str] = None,
     *args,
     **kwargs
-) -> Union[Path, str, onnx.ModelProto]:
+) -> Union[pathlib.Path, str, onnx.ModelProto]:
     """Apply smooth quant."""
     assert calibration_data_reader is not None, "Please provide calibration_data_reader"
     assert isinstance(
-        calibration_data_reader, CalibrationDataReader
-    ), "Please follow onnx_neural_compressor/quantization/calibrate.py to implement calibration_data_reader"
+        calibration_data_reader, data_reader.CalibrationDataReader
+    ), "Please follow onnx_neural_compressor/data_reader.py to implement calibration_data_reader"
 
     # smooth operation
     calibration_data_reader.rewind()
-    smoother = Smoother(
+    smoother = core.Smoother(
         model,
         calibration_data_reader,
         providers=quant_config.providers,
@@ -61,7 +57,7 @@ def smooth_quant_entry(
         # ORT quant API requires str input
         onnx.save_model(
             smoothed_model,
-            Path(tmp_dir).joinpath("smooth.onnx").as_posix(),
+            pathlib.Path(tmp_dir).joinpath("smooth.onnx").as_posix(),
             save_as_external_data=True,
             all_tensors_to_one_file=True,
             location="smooth.onnx_data",
@@ -77,42 +73,44 @@ def smooth_quant_entry(
         quant_config.calibration_data_reader = calibration_data_reader
         quant_config.nodes_to_exclude.extend(excluded_nodes)
         quant_config.convert_to_ort_config()
-        quantize(
-            Path(tmp_dir).joinpath("smooth.onnx").as_posix(),
-            model_output or Path(tmp_dir).joinpath("quant_model.onnx").as_posix(),
+        quantization.quantize(
+            pathlib.Path(tmp_dir).joinpath("smooth.onnx").as_posix(),
+            model_output or pathlib.Path(tmp_dir).joinpath("quant_model.onnx").as_posix(),
             quant_config,
         )
-        model = model_output or onnx.load(Path(tmp_dir).joinpath("quant_model.onnx").as_posix())
+        model = model_output or onnx.load(pathlib.Path(tmp_dir).joinpath("quant_model.onnx").as_posix())
 
     return model
 
 
 ###################### RTN Algo Entry ##################################
-@register_algo(name=RTN)
-def rtn_quantize_entry(model: Union[Path, str], quant_config: RTNConfig, *args, **kwargs) -> onnx.ModelProto:
+@utility.register_algo(name=constants.RTN)
+def rtn_quantize_entry(
+    model: Union[pathlib.Path, str], quant_config: config.RTNConfig, *args, **kwargs
+) -> onnx.ModelProto:
     """The main entry to apply rtn quantization."""
-    from onnx_neural_compressor.algorithms import apply_rtn_on_model
-
     # map config to each op
     model_info = quant_config.get_model_info(model=model)
     configs_mapping = quant_config.to_config_mapping(model_info=model_info)
     logger.debug(configs_mapping)
-    model = apply_rtn_on_model(model, configs_mapping)
+    model = rtn.apply_rtn_on_model(model, configs_mapping)
     return model
 
 
 ###################### GPTQ Algo Entry ##################################
-@register_algo(name=GPTQ)
+@utility.register_algo(name=constants.GPTQ)
 def gptq_quantize_entry(
-    model: Union[Path, str], quant_config: GPTQConfig, calibration_data_reader: CalibrationDataReader, *args, **kwargs
+    model: Union[pathlib.Path, str],
+    quant_config: config.GPTQConfig,
+    calibration_data_reader: data_reader.CalibrationDataReader,
+    *args,
+    **kwargs
 ) -> onnx.ModelProto:
     """The main entry to apply gptq quantization."""
     assert calibration_data_reader is not None, "Please provide calibration_data_reader"
     assert isinstance(
-        calibration_data_reader, CalibrationDataReader
-    ), "Please follow onnx_neural_compressor/quantization/calibrate.py to implement calibration_data_reader"
-
-    from onnx_neural_compressor.algorithms import apply_gptq_on_model
+        calibration_data_reader, data_reader.CalibrationDataReader
+    ), "Please follow onnx_neural_compressor/data_reader.py to implement calibration_data_reader"
 
     # map config to each op
     model_info = quant_config.get_model_info(model=model)
@@ -121,22 +119,24 @@ def gptq_quantize_entry(
 
     # regenerate to ensure data exists
     calibration_data_reader.rewind()
-    model = apply_gptq_on_model(model, configs_mapping, calibration_data_reader)
+    model = gptq.apply_gptq_on_model(model, configs_mapping, calibration_data_reader)
     return model
 
 
 ###################### AWQ Algo Entry ##################################
-@register_algo(name=AWQ)
+@utility.register_algo(name=constants.AWQ)
 def awq_quantize_entry(
-    model: Union[Path, str], quant_config: AWQConfig, calibration_data_reader: CalibrationDataReader, *args, **kwargs
+    model: Union[pathlib.Path, str],
+    quant_config: config.AWQConfig,
+    calibration_data_reader: data_reader.CalibrationDataReader,
+    *args,
+    **kwargs
 ) -> onnx.ModelProto:
     """The main entry to apply awq quantization."""
     assert calibration_data_reader is not None, "Please provide calibration_data_reader"
     assert isinstance(
-        calibration_data_reader, CalibrationDataReader
-    ), "Please follow onnx_neural_compressor/quantization/calibrate.py to implement calibration_data_reader"
-
-    from onnx_neural_compressor.algorithms import apply_awq_on_model
+        calibration_data_reader, data_reader.CalibrationDataReader
+    ), "Please follow onnx_neural_compressor/data_reader.py to implement calibration_data_reader"
 
     # map config to each op
     model_info = quant_config.get_model_info(model=model)
@@ -145,5 +145,5 @@ def awq_quantize_entry(
 
     # regenerate to ensure data exists
     calibration_data_reader.rewind()
-    model = apply_awq_on_model(model, configs_mapping, calibration_data_reader)
+    model = awq.apply_awq_on_model(model, configs_mapping, calibration_data_reader)
     return model
