+++ onnx_neural_compressor/quantization/matmul_4bits_quantizer.py
@@ -12,28 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Union
+from typing import List, Union  # isort: skip
 
-from onnx.onnx_pb import ModelProto
-from onnxruntime.quantization.matmul_4bits_quantizer import WeightOnlyQuantConfig
+import onnx
+from onnx_neural_compressor.quantization import matmul_nbits_quantizer
+from onnxruntime.quantization import matmul_4bits_quantizer
 
-from onnx_neural_compressor.quantization.matmul_nbits_quantizer import (
-    AWQWeightOnlyQuantConfig,
-    GPTQWeightOnlyQuantConfig,
-    MatMulNBitsQuantizer,
-    RTNWeightOnlyQuantConfig,
-)
+RTNWeightOnlyQuantConfig = matmul_nbits_quantizer.RTNWeightOnlyQuantConfig
+AWQWeightOnlyQuantConfig = matmul_nbits_quantizer.AWQWeightOnlyQuantConfig
+GPTQWeightOnlyQuantConfig = matmul_nbits_quantizer.GPTQWeightOnlyQuantConfig
 
 
-class MatMul4BitsQuantizer(MatMulNBitsQuantizer):
+class MatMul4BitsQuantizer(matmul_nbits_quantizer.MatMulNBitsQuantizer):
+
     def __init__(
         self,
-        model: Union[ModelProto, str],
+        model: Union[onnx.ModelProto, str],
         block_size: int = 128,
         is_symmetric: bool = False,
         accuracy_level: int = 0,
         nodes_to_exclude=None,
-        algo_config: WeightOnlyQuantConfig = None,
+        algo_config: matmul_4bits_quantizer.WeightOnlyQuantConfig = None,
         providers: List[str] = ["CPUExecutionProvider"],
     ):
         super().__init__(
