--- onnx_neural_compressor/utils/__init__.py
@@ -1,38 +1,0 @@
-# Copyright (c) 2023 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from onnx_neural_compressor.utils.logger import *
-from onnx_neural_compressor.utils.onnx_model import ONNXModel
-from onnx_neural_compressor.utils.constants import *
-from onnx_neural_compressor.utils.utility import *
-from onnx_neural_compressor.utils.base_config import options
-
-
-__all__ = [
-    "ONNXModel",
-    "CpuInfo",
-    "LazyImport",
-    "singleton",
-    "PRIORITY_RTN",
-    "PRIORITY_GPTQ",
-    "PRIORITY_AWQ",
-    "PRIORITY_SMOOTH_QUANT",
-    "options",
-    "level",
-    "logger",
-    "set_workspace",
-    "set_random_seed",
-    "set_resume_from",
-    "dump_elapsed_time",
-]
