+++ pyproject.toml
@@ -1,8 +1,9 @@
 [tool.isort]
 profile = "black"
 line_length = 120
-known_first_party = ["neural_compressor", "neural_insights", "neural_solution"]
+known_first_party = ["neural_compressor"]
 extend_skip_glob = ["**/__init__.py"]
+force_single_line = true
 
 
 [tool.black]
