+++ test/quantization/test_smooth_quant.py
@@ -21,18 +21,16 @@ import unittest
 
 import numpy as np
 import onnx
+from onnx_neural_compressor import config
+from onnx_neural_compressor import data_reader
+from onnx_neural_compressor.quantization import QuantType
+from onnx_neural_compressor.quantization import algorithm_entry as algos
+from onnx_neural_compressor.quantization import quantize
 from optimum.exporters.onnx import main_export
 
-from onnx_neural_compressor.quantization import (
-    CalibrationDataReader,
-    QuantType,
-    SmoothQuantConfig,
-    get_default_sq_config,
-)
-from onnx_neural_compressor.quantization.quantize import _quantize
 
+class DataReader(data_reader.CalibrationDataReader):
 
-class DataReader(CalibrationDataReader):
     def __init__(self, model):
         model = onnx.load(model)
         batch_size = 1
@@ -60,6 +58,7 @@ class DataReader(CalibrationDataReader):
 
 
 class TestONNXRT3xSmoothQuant(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
         main_export(
@@ -76,65 +75,36 @@ class TestONNXRT3xSmoothQuant(unittest.TestCase):
 
     def test_sq_from_class_beginner(self):
         self.data_reader.rewind()
-        config = get_default_sq_config()
-        model = _quantize(self.gptj, config, self.data_reader)
+        sq_config = config.get_default_sq_config()
+        model = algos.smooth_quant_entry(self.gptj, sq_config, self.data_reader)
         num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
         self.assertEqual(num_muls, 30)
 
     def test_sq_auto_tune_from_class_beginner(self):
         self.data_reader.rewind()
-        config = SmoothQuantConfig(alpha="auto", scales_per_op=False)
-        model = _quantize(self.gptj, config, self.data_reader)
+        sq_config = config.SmoothQuantConfig(alpha="auto", scales_per_op=False)
+        model = algos.smooth_quant_entry(self.gptj, sq_config, self.data_reader)
         num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
         self.assertEqual(num_muls, 15)
-
-    def test_sq_from_dict_beginner(self):
-        config = {
-            "smooth_quant": {
-                "global": {
-                    "alpha": 0.5,
-                    "scales_per_op": False,
-                },
-            }
-        }
-        self.data_reader.rewind()
-        model = _quantize(self.gptj, config, self.data_reader)
-        num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
-        self.assertEqual(num_muls, 15)
-
-    def test_sq_auto_tune_from_dict_beginner(self):
-        config = {
-            "smooth_quant": {
-                "global": {
-                    "alpha": "auto",
-                },
-            }
-        }
-        self.data_reader.rewind()
-        model = _quantize(self.gptj, config, self.data_reader)
-        num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
-        self.assertEqual(num_muls, 30)
 
     def test_sq_ort_param_class_beginner(self):
         self.data_reader.rewind()
-        config = SmoothQuantConfig(weight_type=QuantType.QUInt8, activation_type=QuantType.QUInt8)
-        model = _quantize(self.gptj, config, self.data_reader)
+        sq_config = config.SmoothQuantConfig(weight_type=QuantType.QUInt8, activation_type=QuantType.QUInt8)
+        model = algos.smooth_quant_entry(self.gptj, sq_config, self.data_reader)
         num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
         self.assertTrue(2 in [i.data_type for i in model.graph.initializer])
         self.assertTrue(3 not in [i.data_type for i in model.graph.initializer])
         self.assertEqual(num_muls, 30)
 
     def test_sq_with_ort_like_api(self):
-        from onnx_neural_compressor.quantization import StaticQuantConfig, quantize
-
         self.data_reader.rewind()
-        config = StaticQuantConfig(
+        sq_config = config.StaticQuantConfig(
             self.data_reader,
             weight_type=QuantType.QUInt8,
             activation_type=QuantType.QUInt8,
             extra_options={"SmoothQuant": True, "SmoothQuantAlpha": 0.7, "SmoothQuantCalibIter": 1},
         )
-        quantize(self.gptj, self.quant_gptj, config)
+        quantize(self.gptj, self.quant_gptj, sq_config)
         model = onnx.load(self.quant_gptj)
         num_muls = len([i for i in model.graph.node if i.name.endswith("_smooth_mul") and i.op_type == "Mul"])
         self.assertTrue(2 in [i.data_type for i in model.graph.initializer])
