+++ test/quantization/weight_only/test_awq.py
@@ -1,14 +1,18 @@
 import copy
+import itertools
 import os
 import shutil
 import unittest
 
 import torch
+import transformers
+from onnx_neural_compressor import config
+from onnx_neural_compressor import data_reader
+from onnx_neural_compressor import logger
+from onnx_neural_compressor.quantization import algorithm_entry as algos
+from onnx_neural_compressor.quantization import matmul_4bits_quantizer
+from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 from optimum.exporters.onnx import main_export
-from transformers import AutoTokenizer
-
-from onnx_neural_compressor.quantization.calibrate import CalibrationDataReader
-from onnx_neural_compressor.utils import logger
 
 
 def find_onnx_file(folder_path):
@@ -20,9 +24,10 @@ def find_onnx_file(folder_path):
     return None
 
 
-class DummyNLPDataloader(CalibrationDataReader):
+class DummyNLPDataloader(data_reader.CalibrationDataReader):
+
     def __init__(self, model_name):
-        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
+        self.tokenizer = transformers.AutoTokenizer.from_pretrained(model_name)
         self.sequence_a = "intel-extension-for-transformers is based in SH"
         self.sequence_b = "Where is intel-extension-for-transformers based? NYC or SH"
 
@@ -49,6 +54,7 @@ class DummyNLPDataloader(CalibrationDataReader):
 
 
 class TestAWQQuant(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
         main_export(
@@ -89,17 +95,18 @@ class TestAWQQuant(unittest.TestCase):
 
     def _apply_awq(self, quant_config):
         logger.info(f"Test AWQ with config {quant_config}")
-        from onnx_neural_compressor.quantization.quantize import _quantize
 
         fp32_model = copy.deepcopy(self.gptj)
-        qmodel = _quantize(fp32_model, quant_config, calibration_data_reader=self.calibration_data_reader)
+        qmodel = algos.awq_quantize_entry(
+            fp32_model, quant_config, calibration_data_reader=self.calibration_data_reader
+        )
         self.assertIsNotNone(qmodel)
         return qmodel
 
 
 class TestAWQQuantWithInternalAPI(TestAWQQuant):
+
     def test_awq_params_combination(self):
-        from onnx_neural_compressor.quantization import AWQConfig
 
         # some tests were skipped to accelerate the CI
         # TODO: check params combination.
@@ -114,58 +121,40 @@ class TestAWQQuantWithInternalAPI(TestAWQQuant):
             "enable_auto_scale": [True, False],
             "enable_mse_search": [True, False],
         }
-        from itertools import product
 
-        keys = AWQConfig.params_list
-        for value in product(*awq_options.values()):
+        keys = config.AWQConfig.params_list
+        for value in itertools.product(*awq_options.values()):
             d = dict(zip(keys, value))
             print(d)
-            quant_config = AWQConfig(**d)
+            quant_config = config.AWQConfig(**d)
             qmodel = self._apply_awq(quant_config)
             self.assertEqual(self._count_woq_matmul(qmodel, bits=value[1], group_size=value[2]), 30)
 
     def test_awq_config(self):
-        from onnx_neural_compressor.quantization import AWQConfig
 
-        awq_config1 = AWQConfig(weight_bits=4)
+        awq_config1 = config.AWQConfig(weight_bits=4)
         quant_config_dict = {
             "awq": {"weight_bits": 4},
         }
-        awq_config2 = AWQConfig.from_dict(quant_config_dict["awq"])
+        awq_config2 = config.AWQConfig.from_dict(quant_config_dict["awq"])
         self.assertEqual(awq_config1.to_dict(), awq_config2.to_dict())
 
     def test_quantize_awq_from_dict_default(self):
-        from onnx_neural_compressor.quantization import get_default_awq_config
 
-        qmodel = self._apply_awq(quant_config=get_default_awq_config())
-        self.assertIsNotNone(qmodel)
-        self.assertTrue(self._check_model_is_quantized(qmodel))
-
-    def test_quantize_awq_from_dict_beginner(self):
-        quant_config = {
-            "awq": {
-                "weight_bits": 4,
-                "weight_group_size": 32,
-            },
-        }
-        qmodel = self._apply_awq(quant_config)
-        self.assertIsNotNone(qmodel)
+        qmodel = self._apply_awq(quant_config=config.get_default_awq_config())
         self.assertIsNotNone(qmodel)
         self.assertTrue(self._check_model_is_quantized(qmodel))
 
     def test_quantize_awq_from_class_beginner(self):
-        from onnx_neural_compressor.quantization import AWQConfig
 
-        quant_config = AWQConfig(weight_bits=4, weight_group_size=32)
+        quant_config = config.AWQConfig(weight_bits=4, weight_group_size=32)
         qmodel = self._apply_awq(quant_config)
         self.assertIsNotNone(qmodel)
 
     def test_quantize_awq_fallback_from_class_beginner(self):
-        from onnx_neural_compressor.quantization import AWQConfig
 
-        fp32_config = AWQConfig(weight_dtype="fp32")
-        quant_config = AWQConfig(
-            weight_bits=4,
+        fp32_config = config.AWQConfig(weight_dtype="fp32")
+        quant_config = config.AWQConfig(
             weight_dtype="int",
             weight_sym=False,
             weight_group_size=32,
@@ -176,49 +165,10 @@ class TestAWQQuantWithInternalAPI(TestAWQQuant):
         self.assertEqual(self._count_woq_matmul(qmodel), 29)
         self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
-    def test_quantize_awq_from_dict_advance(self):
-        quant_config = {
-            "awq": {
-                "global": {
-                    "weight_bits": 4,
-                    "weight_group_size": 32,
-                },
-                "local": {
-                    "/h.4/mlp/fc_out/MatMul": {
-                        "weight_dtype": "fp32",
-                    }
-                },
-            }
-        }
-        qmodel = self._apply_awq(quant_config)
-        self.assertIsNotNone(qmodel)
-        self.assertEqual(self._count_woq_matmul(qmodel), 29)
-        self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
-
-        quant_config = {
-            "awq": {
-                "global": {
-                    "weight_bits": 4,
-                    "weight_group_size": 32,
-                },
-                "local": {
-                    "/h.4/mlp/fc_out/MatMul": {
-                        "weight_bits": 8,
-                        "weight_group_size": 32,
-                    }
-                },
-            }
-        }
-        qmodel = self._apply_awq(quant_config)
-        self.assertIsNotNone(qmodel)
-        for node in qmodel.graph.node:
-            if node.name == "/h.4/mlp/fc_out/MatMul":
-                self.assertTrue(node.input[1].endswith("Q8G32"))
-
 
 class TestAWQQuantWithORTLikeAPI(TestAWQQuant):
+
     def test_awq_config_4bits(self):
-        from onnx_neural_compressor.quantization import matmul_4bits_quantizer
 
         algo_config = matmul_4bits_quantizer.AWQWeightOnlyQuantConfig(
             calibration_data_reader=self.calibration_data_reader
@@ -235,7 +185,6 @@ class TestAWQQuantWithORTLikeAPI(TestAWQQuant):
         self.assertTrue(self._check_model_is_quantized(quant.model))
 
     def test_awq_config_4bits_with_exclude_node(self):
-        from onnx_neural_compressor.quantization import matmul_4bits_quantizer
 
         algo_config = matmul_4bits_quantizer.AWQWeightOnlyQuantConfig(
             calibration_data_reader=self.calibration_data_reader
@@ -254,7 +203,6 @@ class TestAWQQuantWithORTLikeAPI(TestAWQQuant):
         self.assertFalse(self._check_node_is_quantized(quant.model, "/h.4/mlp/fc_out/MatMul"))
 
     def test_awq_config_nbits(self):
-        from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 
         algo_config = matmul_nbits_quantizer.AWQWeightOnlyQuantConfig(
             calibration_data_reader=self.calibration_data_reader
@@ -273,7 +221,6 @@ class TestAWQQuantWithORTLikeAPI(TestAWQQuant):
             self.assertEqual(self._count_woq_matmul(quant.model, bits=n_bits, group_size=32), 30)
 
     def test_awq_config_nbits_with_exclude_node(self):
-        from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 
         algo_config = matmul_nbits_quantizer.AWQWeightOnlyQuantConfig(
             calibration_data_reader=self.calibration_data_reader
