+++ test/quantization/weight_only/test_rtn.py
@@ -1,11 +1,15 @@
 import copy
+import itertools
 import os
 import shutil
 import unittest
 
+from onnx_neural_compressor import config
+from onnx_neural_compressor import logger
+from onnx_neural_compressor.quantization import algorithm_entry as algos
+from onnx_neural_compressor.quantization import matmul_4bits_quantizer
+from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 from optimum.exporters.onnx import main_export
-
-from onnx_neural_compressor.utils import logger
 
 
 def find_onnx_file(folder_path):
@@ -18,6 +22,7 @@ def find_onnx_file(folder_path):
 
 
 class TestRTNQuant(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
         main_export(
@@ -57,17 +62,16 @@ class TestRTNQuant(unittest.TestCase):
 
     def _apply_rtn(self, quant_config):
         logger.info(f"Test RTN with config {quant_config}")
-        from onnx_neural_compressor.quantization.quantize import _quantize
 
         fp32_model = copy.deepcopy(self.gptj)
-        qmodel = _quantize(fp32_model, quant_config)
+        qmodel = algos.rtn_quantize_entry(fp32_model, quant_config)
         self.assertIsNotNone(qmodel)
         return qmodel
 
 
 class TestRTNQuantWithInternalAPI(TestRTNQuant):
+
     def test_rtn_params_combination(self):
-        from onnx_neural_compressor.quantization import RTNConfig
 
         # some tests were skipped to accelerate the CI
         # TODO: check params combination.
@@ -79,61 +83,39 @@ class TestRTNQuantWithInternalAPI(TestRTNQuant):
             "weight_sym": [True, False],
             "act_dtype": ["fp32"],
         }
-        from itertools import product
 
-        keys = RTNConfig.params_list
-        for value in product(*rtn_options.values()):
+        keys = config.RTNConfig.params_list
+        for value in itertools.product(*rtn_options.values()):
             d = dict(zip(keys, value))
-            quant_config = RTNConfig(**d)
+            quant_config = config.RTNConfig(**d)
             qmodel = self._apply_rtn(quant_config)
             self.assertEqual(self._count_woq_matmul(qmodel, bits=value[1], group_size=value[2]), 30)
 
     def test_rtn_config(self):
-        from onnx_neural_compressor.quantization import RTNConfig
 
-        rtn_config1 = RTNConfig(weight_bits=4)
+        rtn_config1 = config.RTNConfig(weight_bits=4)
         quant_config_dict = {
             "rtn": {"weight_bits": 4},
         }
-        rtn_config2 = RTNConfig.from_dict(quant_config_dict["rtn"])
+        rtn_config2 = config.RTNConfig.from_dict(quant_config_dict["rtn"])
         self.assertEqual(rtn_config1.to_dict(), rtn_config2.to_dict())
 
     def test_quantize_rtn_from_dict_default(self):
-        from onnx_neural_compressor.quantization import get_default_rtn_config
-        from onnx_neural_compressor.quantization.quantize import _quantize
 
-        qmodel = self._apply_rtn(quant_config=get_default_rtn_config())
-        self.assertIsNotNone(qmodel)
-        self.assertTrue(self._check_model_is_quantized(qmodel))
-
-    def test_quantize_rtn_from_dict_beginner(self):
-        from onnx_neural_compressor.quantization.quantize import _quantize
-
-        quant_config = {
-            "rtn": {
-                "weight_bits": 4,
-                "weight_group_size": 32,
-            },
-        }
-        qmodel = self._apply_rtn(quant_config)
-        self.assertIsNotNone(qmodel)
+        qmodel = self._apply_rtn(quant_config=config.get_default_rtn_config())
         self.assertIsNotNone(qmodel)
         self.assertTrue(self._check_model_is_quantized(qmodel))
 
     def test_quantize_rtn_from_class_beginner(self):
-        from onnx_neural_compressor.quantization import RTNConfig
-        from onnx_neural_compressor.quantization.quantize import _quantize
 
-        quant_config = RTNConfig(weight_bits=4, weight_group_size=32)
+        quant_config = config.RTNConfig(weight_bits=4, weight_group_size=32)
         qmodel = self._apply_rtn(quant_config)
         self.assertIsNotNone(qmodel)
 
     def test_quantize_rtn_fallback_from_class_beginner(self):
-        from onnx_neural_compressor.quantization import RTNConfig
-        from onnx_neural_compressor.quantization.quantize import _quantize
 
-        fp32_config = RTNConfig(weight_dtype="fp32")
-        quant_config = RTNConfig(
+        fp32_config = config.RTNConfig(weight_dtype="fp32")
+        quant_config = config.RTNConfig(
             weight_bits=4,
             weight_dtype="int",
             weight_sym=False,
@@ -145,51 +127,10 @@ class TestRTNQuantWithInternalAPI(TestRTNQuant):
         self.assertEqual(self._count_woq_matmul(qmodel), 29)
         self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
-    def test_quantize_rtn_from_dict_advance(self):
-        from onnx_neural_compressor.quantization.quantize import _quantize
-
-        quant_config = {
-            "rtn": {
-                "global": {
-                    "weight_bits": 4,
-                    "weight_group_size": 32,
-                },
-                "local": {
-                    "/h.4/mlp/fc_out/MatMul": {
-                        "weight_dtype": "fp32",
-                    }
-                },
-            }
-        }
-        qmodel = self._apply_rtn(quant_config)
-        self.assertIsNotNone(qmodel)
-        self.assertEqual(self._count_woq_matmul(qmodel), 29)
-        self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
-
-        quant_config = {
-            "rtn": {
-                "global": {
-                    "weight_bits": 4,
-                    "weight_group_size": 32,
-                },
-                "local": {
-                    "/h.4/mlp/fc_out/MatMul": {
-                        "weight_bits": 8,
-                        "weight_group_size": 32,
-                    }
-                },
-            }
-        }
-        qmodel = self._apply_rtn(quant_config)
-        self.assertIsNotNone(qmodel)
-        for node in qmodel.graph.node:
-            if node.name == "/h.4/mlp/fc_out/MatMul":
-                self.assertTrue(node.input[1].endswith("Q8G32"))
-
 
 class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
+
     def test_rtn_config_4bits(self):
-        from onnx_neural_compressor.quantization import matmul_4bits_quantizer
 
         algo_config = matmul_4bits_quantizer.RTNWeightOnlyQuantConfig()
 
@@ -204,7 +145,6 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
         self.assertTrue(self._check_model_is_quantized(quant.model))
 
     def test_rtn_config_4bits_with_exclude_node(self):
-        from onnx_neural_compressor.quantization import matmul_4bits_quantizer
 
         algo_config = matmul_4bits_quantizer.RTNWeightOnlyQuantConfig()
 
@@ -221,7 +161,6 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
         self.assertFalse(self._check_node_is_quantized(quant.model, "/h.4/mlp/fc_out/MatMul"))
 
     def test_rtn_config_nbits(self):
-        from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 
         algo_config = matmul_nbits_quantizer.RTNWeightOnlyQuantConfig()
 
@@ -238,7 +177,6 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
             self.assertEqual(self._count_woq_matmul(quant.model, bits=n_bits, group_size=32), 30)
 
     def test_rtn_config_nbits_with_exclude_node(self):
-        from onnx_neural_compressor.quantization import matmul_nbits_quantizer
 
         algo_config = matmul_nbits_quantizer.RTNWeightOnlyQuantConfig()
