+++ test/utils/test_general.py
@@ -1,19 +1,14 @@
-"""Tests for utils components."""
+"""Tests for general components."""
 
 import unittest
-from typing import Any, List, Optional, Tuple, Union
 
-from onnx_neural_compressor.utils import logger
-from onnx_neural_compressor.utils.base_config import (
-    BaseConfig,
-    config_registry,
-    get_all_config_set_from_config_registry,
-    register_config,
-    register_supported_configs,
-)
-from onnx_neural_compressor.utils.base_tuning import ConfigLoader, ConfigSet, Evaluator, SequentialSampler
-from onnx_neural_compressor.utils.constants import DEFAULT_WHITE_LIST, OP_NAME_OR_MODULE_TYPE
-from onnx_neural_compressor.utils.tuning_param import TuningParam
+from onnx_neural_compressor import config
+from onnx_neural_compressor import constants
+from onnx_neural_compressor import logger
+from onnx_neural_compressor.quantization import tuning
+
+from typing import Any, Callable, List, Optional, Tuple, Union  # isort: skip
+
 
 PRIORITY_FAKE_ALGO = 100
 FAKE_CONFIG_NAME = "fake"
@@ -21,11 +16,11 @@ PRIORITY_FAKE_ALGO_1 = 90
 FAKE_CONFIG_NAME_1 = "fake_one"
 DEFAULT_WEIGHT_BITS = [4, 6]
 
-
 FAKE_MODEL_INFO = [("OP1_NAME", "OP_TYPE1"), ("OP2_NAME", "OP_TYPE1"), ("OP3_NAME", "OP_TYPE2")]
 
 
 class FakeModel:
+
     def __init__(self) -> None:
         self.name = "fake_model"
 
@@ -36,15 +31,15 @@ class FakeModel:
         return "FakeModel"
 
 
-@register_config(algo_name=FAKE_CONFIG_NAME, priority=PRIORITY_FAKE_ALGO)
-class FakeAlgoConfig(BaseConfig):
+@config.register_config(algo_name=FAKE_CONFIG_NAME, priority=PRIORITY_FAKE_ALGO)
+class FakeAlgoConfig(config.BaseConfig):
     """Config class for fake algo."""
 
     supported_configs: List = []
     params_list = [
         "weight_dtype",
         "weight_bits",
-        TuningParam("target_op_type_list", tunable_type=List[List[str]]),
+        config.TuningParam("target_op_type_list", tunable_type=List[List[str]]),
     ]
     name = FAKE_CONFIG_NAME
 
@@ -53,7 +48,7 @@ class FakeAlgoConfig(BaseConfig):
         weight_dtype: str = "int",
         weight_bits: int = 4,
         target_op_type_list: List[str] = ["Conv", "Gemm"],
-        white_list: Optional[List[OP_NAME_OR_MODULE_TYPE]] = DEFAULT_WHITE_LIST,
+        white_list: Optional[List[Union[str, Callable]]] = constants.DEFAULT_WHITE_LIST,
     ):
         """Init fake config.
 
@@ -96,15 +91,15 @@ def get_default_fake_config() -> FakeAlgoConfig:
     return FakeAlgoConfig()
 
 
-@register_config(algo_name=FAKE_CONFIG_NAME_1, priority=PRIORITY_FAKE_ALGO_1)
-class FakeAlgoOneConfig(BaseConfig):
+@config.register_config(algo_name=FAKE_CONFIG_NAME_1, priority=PRIORITY_FAKE_ALGO_1)
+class FakeAlgoOneConfig(config.BaseConfig):
     """Config class for fake algo."""
 
     supported_configs: List = []
     params_list = [
         "weight_dtype",
         "weight_bits",
-        TuningParam("target_op_type_list", tunable_type=List[List[str]]),
+        config.TuningParam("target_op_type_list", tunable_type=List[List[str]]),
     ]
     name = FAKE_CONFIG_NAME_1
 
@@ -113,7 +108,7 @@ class FakeAlgoOneConfig(BaseConfig):
         weight_dtype: str = "int",
         weight_bits: int = 4,
         target_op_type_list: List[str] = ["Conv", "Gemm"],
-        white_list: Optional[List[OP_NAME_OR_MODULE_TYPE]] = DEFAULT_WHITE_LIST,
+        white_list: Optional[List[Union[str, Callable]]] = constants.DEFAULT_WHITE_LIST,
     ):
         """Init fake config.
 
@@ -147,19 +142,21 @@ class FakeAlgoOneConfig(BaseConfig):
         return FakeAlgoOneConfig(weight_bits=DEFAULT_WEIGHT_BITS)
 
 
-def get_all_config_set() -> Union[BaseConfig, List[BaseConfig]]:
-    return get_all_config_set_from_config_registry()
+def get_all_config_set() -> Union[config.BaseConfig, List[config.BaseConfig]]:
+    return config.get_all_config_set_from_config_registry()
 
 
-register_supported_configs()
+config.register_supported_configs()
 
 
 class TestEvaluator(unittest.TestCase):
+
     def test_single_eval_fn(self):
+
         def fake_eval_fn(model):
             return 1.0
 
-        evaluator = Evaluator()
+        evaluator = tuning.Evaluator()
         evaluator.set_eval_fn_registry(fake_eval_fn)
         evaluator.self_check()
         self.assertEqual(evaluator.get_number_of_eval_functions(), 1)
@@ -172,13 +169,14 @@ class TestEvaluator(unittest.TestCase):
 
         eval_fns = {"eval_fn": eval_acc_fn, "weight": 0.5, "name": "accuracy"}
 
-        evaluator = Evaluator()
+        evaluator = tuning.Evaluator()
         evaluator.set_eval_fn_registry(eval_fns)
         evaluator.self_check()
         self.assertEqual(evaluator.get_number_of_eval_functions(), 1)
 
 
 class TestBaseConfig(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
         pass
@@ -195,7 +193,7 @@ class TestBaseConfig(unittest.TestCase):
         fake_default_config = get_default_fake_config()
         self.assertEqual(fake_default_config.weight_dtype, "int")
         config_set = get_all_config_set()
-        self.assertEqual(len(config_set), len(config_registry.get_all_config_cls()))
+        self.assertEqual(len(config_set), len(config.config_registry.get_all_config_cls()))
         self.assertEqual([i for i in config_set if i.name == FAKE_CONFIG_NAME][0].weight_bits, DEFAULT_WEIGHT_BITS)
 
     def test_config_expand_complex_tunable_type(self):
@@ -220,9 +218,10 @@ class TestBaseConfig(unittest.TestCase):
 
 
 class TestConfigSet(unittest.TestCase):
+
     def setUp(self):
         self.config_set = [get_default_fake_config(), get_default_fake_config()]
-        self.config_set_obj = ConfigSet.from_fwk_configs(self.config_set)
+        self.config_set_obj = tuning.ConfigSet.from_fwk_configs(self.config_set)
 
     def test_config_set(self) -> None:
         self.assertEqual(len(self.config_set_obj), len(self.config_set))
@@ -230,23 +229,25 @@ class TestConfigSet(unittest.TestCase):
 
 
 class TestConfigSampler(unittest.TestCase):
+
     def setUp(self):
         self.config_set = [get_default_fake_config(), get_default_fake_config()]
-        self.seq_sampler = SequentialSampler(self.config_set)
+        self.seq_sampler = tuning.SequentialSampler(self.config_set)
 
     def test_config_sampler(self) -> None:
         self.assertEqual(list(self.seq_sampler), list(range(len(self.config_set))))
 
 
 class TestConfigLoader(unittest.TestCase):
+
     def setUp(self):
         self.config_set = [FakeAlgoConfig(weight_bits=4), FakeAlgoConfig(weight_bits=8)]
-        self.loader = ConfigLoader(self.config_set)
+        self.loader = tuning.ConfigLoader(self.config_set)
 
     def test_config_loader(self) -> None:
         self.assertEqual(len(list(self.loader)), len(self.config_set))
-        for i, config in enumerate(self.loader):
-            self.assertEqual(config, self.config_set[i])
+        for i, cfg in enumerate(self.loader):
+            self.assertEqual(cfg, self.config_set[i])
 
 
 if __name__ == "__main__":
